Fixing free turning into main

+++ hud.h
@@ -19,6 +19,9 @@ typedef struct {
     unsigned short height_lives;
 } hud;
 
+// Cria a hud
 hud* create_hud();
-void generate_hud(hud* hud, player* player, int round, ALLEGRO_BITMAP* sprite_sheet, ALLEGRO_FONT* font);
+// Gera a hud na tela
+void generate_hud(hud* hud, player* player, unsigned short round, ALLEGRO_BITMAP* sprite_sheet, ALLEGRO_FONT* font);
+
 #endif // __HUD__
